Apply tabs to tx calldata
